+++ src/pages/Appearance/Programs/ProgramGlobalSettings.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from '@dhis2/prop-types'
 import i18n from '@dhis2/d2-i18n'
-import Wrapper from '../../../components/wrapper'
+import Wrapper from '../../../components/Wrapper'
 import PageSubtitle from '../../../components/page/PageSubtitle'
 import { TableHeader } from '../../../components/table'
 import { GlobalProgramCompletion } from '../../../components/field'
