+++ src/pages/Synchronization/Datasets/SpecificSettings.js
@@ -3,7 +3,7 @@ import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
 import { Divider } from '@dhis2/ui'
 import TableRow from '../../../components/settingsTable/TableRow'
-import Wrapper from '../../../components/wrapper'
+import Wrapper from '../../../components/Wrapper'
 import { InputNumber } from '../../../components/inputs'
 import { DataSpecificSetting } from '../../../constants/data-set-settings'
 import { getPeriodDefaultValueByType } from './helper'
